scipt and code chnages

+++ setup/configure.sh
@@ -4,9 +4,34 @@ set -ex
 EASY_RSA_LOC="/etc/openvpn/easyrsa"
 SERVER_CERT="${EASY_RSA_LOC}/pki/issued/server.crt"
 
-OVPN_SRV_NET=${OVPN_SERVER_NET:-192.168.100.0}
+OVPN_SRV_NET=${OVPN_SERVER_NET:-10.10.0.0}
 OVPN_SRV_MASK=${OVPN_SERVER_MASK:-255.255.255.0}
 OVPN_PASSWD_AUTH=false
+
+set -e
+
+TARGETARCH=$(dpkg --print-architecture) 
+
+sudo apt update -y
+
+sudo apt install -y openvpn iptables
+
+if [ ! -f "/usr/local/bin/easyrsa" ]; then
+  sudo apt install easy-rsa
+  sudo ln -sf /usr/share/easy-rsa/easyrsa /usr/local/bin/easyrsa
+fi
+
+if [ ! -f "/usr/local/bin/openvpn-user" ]; then
+  cd /tmp
+  wget "https://github.com/pashcovich/openvpn-user/releases/download/v1.0.4/openvpn-user-linux-${TARGETARCH}.tar.gz" -O - | sudo tar xz -C /usr/local/bin
+fi
+
+if [ -f "/usr/local/bin/openvpn-user-${TARGETARCH}" ]; then
+  sudo ln -sf /usr/local/bin/openvpn-user-${TARGETARCH} /usr/local/bin/openvpn-user
+fi
+
+mkdir -p /etc/openvpn/easyrsa
+mkdir -p /etc/openvpn/ccd
 
 cd $EASY_RSA_LOC
 
@@ -39,7 +64,7 @@ if [ ! -c /dev/net/tun ]; then
     mknod /dev/net/tun c 10 200
 fi
 
-cp -f /etc/openvpn/setup/openvpn.conf /etc/openvpn/openvpn.conf
+cp -f /home/ubuntu/open-vpn-2fa/setup/openvpn.conf /etc/openvpn/openvpn.conf
 
 if [ ${OVPN_PASSWD_AUTH} = "true" ]; then
   mkdir -p /etc/openvpn/scripts/
@@ -54,11 +79,9 @@ fi
 [ -d $EASY_RSA_LOC/pki ] && chmod 755 $EASY_RSA_LOC/pki
 [ -f $EASY_RSA_LOC/pki/crl.pem ] && chmod 644 $EASY_RSA_LOC/pki/crl.pem
 
-sudo mkdir -p /etc/openvpn/ccd
 
-sudo chown -R root /etc/google-auth
-
-apt update && apt install -y \
+if [ ! -f "/usr/local/lib/security/pam_google_authenticator.so" ]; then
+  apt update && apt install -y \
   build-essential \
   linux-headers-$(uname -r) \
   autoconf \
@@ -69,40 +92,47 @@ apt update && apt install -y \
   git \
   libpam0g-dev \
   libpam-google-authenticator \
-  libqrencode-dev
-cd /tmp
-rm -rf google-authenticator-libpam
-git clone https://github.com/google/google-authenticator-libpam
-cd google-authenticator-libpam/
-./bootstrap.sh
-./configure
-make
-make install
+  qrencode
 
-bash -c 'cat > /etc/pam.d/openvpn <<EOF
-auth    requisite       /usr/local/lib/security/pam_google_authenticator.so secret=/etc/google-auth/${USER}  user=root
-account    required     pam_permit.so
-EOF'
+  cd /tmp
+  rm -rf google-authenticator-libpam
+  git clone https://github.com/google/google-authenticator-libpam
+  cd google-authenticator-libpam/
+  ./bootstrap.sh
+  ./configure
+  make
+  make install
+  rm -rf google-authenticator-libpam
+fi
 
-bash -c 'cat > /etc/openvpn/google-auth.sh <<EOF
-#!/bin/bash
+if [ ! -f "/etc/pam.d/openvpn" ]; then
+  bash -c 'cat > /etc/pam.d/openvpn <<EOF
+  auth    requisite       /usr/local/lib/security/pam_google_authenticator.so secret=/etc/google-auth/\${USER}  user=root
+  account    required     pam_permit.so
+  EOF'
+fi
 
-CLIENT=$1
-HOST=$(hostname)
-R="\e[0;91m"
-G="\e[0;92m"
-W="\e[0;97m"
-B="\e[1m"
-C="\e[0m"
+if [ ! -f "/etc/openvpn/google-auth.sh" ]; then
+  sudo mkdir -p /etc/google-auth
+  sudo chown -R root /etc/google-auth
 
-google-authenticator -t -d -f -r 3 -R 30 -W -C -s "/etc/google-auth/${CLIENT}" || { echo -e "${R}${B}error generating QR code${C}"; exit 1; }
-secret=$(head -n 1 "/etc/google-auth/${CLIENT}")
-qrencode -t PNG -o "/etc/google-auth/${CLIENT}.png" "otpauth://totp/${CLIENT}@${HOST}?secret=${secret}&issuer=openvpn" || { echo -e "${R}${B}Error generating PNG${C}"; exit 1; }
-EOF'
+  sudo bash -c 'cat > /etc/openvpn/google-auth.sh <<EOF
+  #!/bin/bash
+  
+  CLIENT=\$1
+  HOST=\$(hostname)
+  R="\e[0;91m"
+  G="\e[0;92m"
+  W="\e[0;97m"
+  B="\e[1m"
+  C="\e[0m"
+  
+  google-authenticator -t -d -f -r 3 -R 30 -W -C -s "/etc/google-auth/\${CLIENT}" || { echo -e "\${R}\${B}error generating QR code\${C}"; exit 1; }
+  secret=\$(head -n 1 "/etc/google-auth/\${CLIENT}")
+  qrencode -t PNG -o "/etc/google-auth/\${CLIENT}.png" "otpauth://totp/\${CLIENT}@\${HOST}?secret=\${secret}&issuer=openvpn" || { echo -e "\${R}\${B}Error generating PNG\${C}"; exit 1; }
+  EOF'
+  
+  sudo chmod +x /etc/openvpn/google-auth.sh
+fi
 
-sudo chmod +x /etc/openvpn/google-auth.sh
-sudo chmod +x ./bootstrap.sh
-sudo chmod +x ./build.sh
-
-
-openvpn --config /etc/openvpn/openvpn.conf --client-config-dir /etc/openvpn/ccd --port 1194 --proto tcp --management 127.0.0.1 8989 --dev tun --server ${OVPN_SRV_NET} ${OVPN_SRV_MASK}
+openvpn --config /etc/openvpn/openvpn.conf
